refactor: useChat split into hooks/chat module
