+++ packages/shared/src/hooks/chat/index.ts
@@ -1,0 +1,2 @@
+export * from './useChat';
+export * from './types';
